+++ circuit_maintenance_parser/parser.py
@@ -214,3 +214,29 @@ class EmailSubjectParser(Parser):
     def bytes_to_string(string):
         """Convert bytes variable to a string."""
         return string.decode("utf-8")
+
+
+class Csv(Parser):
+    """Csv parser."""
+
+    _data_types = ["application/csv", "text/csv", "application/octet-stream"]
+
+    def parse(self, raw: bytes) -> List[Dict]:
+        """Execute parsing."""
+        result = []
+
+        data_base: Dict[str, Union[int, str, Iterable]] = {}
+        try:
+            for data in self.parse_csv(raw, data_base):
+                result.append(data)
+            logger.debug("Successful parsing for %s", self.__class__.__name__)
+
+            return result
+
+        except Exception as exc:
+            raise ParserError from exc
+
+    @staticmethod
+    def parse_csv(raw: bytes, data_base: Dict) -> List[Dict]:
+        """Custom CSV parsing."""
+        raise NotImplementedError