+++ circuit_maintenance_parser/provider.py
@@ -15,9 +15,11 @@ from circuit_maintenance_parser.errors import ProcessorError, ProviderError
 from circuit_maintenance_parser.processor import CombinedProcessor, SimpleProcessor, GenericProcessor
 
 from circuit_maintenance_parser.parsers.cogent import HtmlParserCogent1
+from circuit_maintenance_parser.parsers.colt import ICalParserColt1, CsvParserColt1
 from circuit_maintenance_parser.parsers.gtt import HtmlParserGTT1
 from circuit_maintenance_parser.parsers.lumen import HtmlParserLumen1
 from circuit_maintenance_parser.parsers.megaport import HtmlParserMegaport1
+from circuit_maintenance_parser.parsers.momentum import HtmlParserMomentum1, SubjectParserMomentum1
 from circuit_maintenance_parser.parsers.seaborn import (
     HtmlParserSeaborn1,
     HtmlParserSeaborn2,
@@ -111,6 +113,15 @@ class Cogent(GenericProvider):
     _default_organizer = "support@cogentco.com"
 
 
+class Colt(GenericProvider):
+    """Cogent provider custom class."""
+
+    _processors: List[GenericProcessor] = [
+        CombinedProcessor(data_parsers=[ICalParserColt1, CsvParserColt1]),
+    ]
+    _default_organizer = "PlannedWorks@colt.net"
+
+
 class EUNetworks(GenericProvider):
     """EUNetworks provider custom class."""
 
@@ -142,6 +153,15 @@ class Megaport(GenericProvider):
         CombinedProcessor(data_parsers=[EmailDateParser, HtmlParserMegaport1]),
     ]
     _default_organizer = "support@megaport.com"
+
+
+class Momentum(GenericProvider):
+    """Momentum provider custom class."""
+
+    _processors: List[GenericProcessor] = [
+        CombinedProcessor(data_parsers=[EmailDateParser, HtmlParserMomentum1, SubjectParserMomentum1]),
+    ]
+    _default_organizer = "maintenance@momentumtelecom.com"
 
 
 class NTT(GenericProvider):