+++ tests/unit/test_parsers.py
@@ -7,9 +7,11 @@ import pytest
 
 from circuit_maintenance_parser.parser import ICal, EmailDateParser
 from circuit_maintenance_parser.parsers.cogent import HtmlParserCogent1
+from circuit_maintenance_parser.parsers.colt import ICalParserColt1, CsvParserColt1
 from circuit_maintenance_parser.parsers.gtt import HtmlParserGTT1
 from circuit_maintenance_parser.parsers.lumen import HtmlParserLumen1
 from circuit_maintenance_parser.parsers.megaport import HtmlParserMegaport1
+from circuit_maintenance_parser.parsers.momentum import HtmlParserMomentum1
 from circuit_maintenance_parser.parsers.seaborn import (
     HtmlParserSeaborn1,
     HtmlParserSeaborn2,
@@ -44,6 +46,17 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             HtmlParserCogent1,
             Path(dir_path, "data", "cogent", "cogent2.html"),
             Path(dir_path, "data", "cogent", "cogent2_result.json"),
+        ),
+        # Colt
+        (
+            ICalParserColt1,
+            Path(dir_path, "data", "colt", "colt1"),
+            Path(dir_path, "data", "colt", "colt1_result.json"),
+        ),
+        (
+            CsvParserColt1,
+            Path(dir_path, "data", "colt", "colt2.csv"),
+            Path(dir_path, "data", "colt", "colt2_result.json"),
         ),
         # GTT
         (
@@ -92,6 +105,12 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             HtmlParserMegaport1,
             Path(dir_path, "data", "megaport", "megaport2.html"),
             Path(dir_path, "data", "megaport", "megaport2_result.json"),
+        ),
+        # Momentum
+        (
+            HtmlParserMomentum1,
+            Path(dir_path, "data", "momentum", "momentum1.eml"),
+            Path(dir_path, "data", "momentum", "momentum1_html_parser_result.json"),
         ),
         # NTT
         (ICal, Path(dir_path, "data", "ntt", "ntt1"), Path(dir_path, "data", "ntt", "ntt1_result.json"),),